+++ src/com/intellij/idea/plugin/hybris/settings/HybrisProjectSettings.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -20,6 +20,7 @@ package com.intellij.idea.plugin.hybris.settings
 
 import com.intellij.idea.plugin.hybris.facet.ExtensionDescriptor
 import com.intellij.idea.plugin.hybris.flexibleSearch.settings.FlexibleSearchSettings
+import com.intellij.idea.plugin.hybris.groovy.settings.GroovySettings
 import com.intellij.idea.plugin.hybris.impex.settings.ImpexSettings
 import com.intellij.idea.plugin.hybris.polyglotQuery.settings.PolyglotQuerySettings
 import java.util.*
@@ -28,6 +29,7 @@ data class HybrisProjectSettings(
     var flexibleSearchSettings: FlexibleSearchSettings = FlexibleSearchSettings(),
     var polyglotQuerySettings: PolyglotQuerySettings = PolyglotQuerySettings(),
     var impexSettings: ImpexSettings = ImpexSettings(),
+    var groovySettings: GroovySettings = GroovySettings(),
 
     var customDirectory: String? = null,
     var hybrisDirectory: String? = null,